Make the MISSINGNO glitch display correctly
Got MISSINGNO to work! All I needed to implement was the option to
specify a size to render (not the same as the size in the binary) and
fix the position adjustment calculation to reproduce the overflow.
Also make a few improvements to the main call so that reading straight
from a given offset in the ROM is possible.